Updated
Updated to fix the warning detected by Pytest. It seems that using np.sum() on a generator function will be deprecated soon due to some performance issue. So computation of corr array is split into two parts: get the list of values with list comprehension, which is then converted to a numpy.ndarray before summed by np.sum().